Remove OverdueListItem sealed class
